+++ src/Generator/Internal.php
@@ -18,7 +18,7 @@ class Internal extends AbstractGenerator implements NumberGenerator
 {
     public function isSupported()
     {
-        return version_compare(PHP_VERSION, '7.0', '>=');
+        return version_compare(PHP_VERSION, '7.0') >= 0;
     }
 
     protected function readBytes($count)