+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Changelog #
 
+## v1.3.1 (2017-09-09) ##
+
+  * The `ByteNumberGenerator::getNumber()` will now throw an `GeneratorException`
+    if the difference between minimum and maximum is not an integer
+  * Rely on `unpack()` rather than `hexdec(bin2hex())` due to being less likely
+    to be affected by timing vulnerabilities.
+  * Require phpcs and php-cs-fixer as external dependencies in the travis build
+    instead of including them as dev dependencies
+
 ## v1.3.0 (2017-07-14) ##
 
   * The minimum required PHP version has been increased to 5.6